cyy/requirements.txt
streamlit
Pillow
stepic
cryptography
requests
streamlit-lottie